refactor: move MembershipStatus to own file
